+++ projection.hpp
@@ -1,3 +1,6 @@
+#ifndef PROJECTION_HPP
+#define PROJECTION_HPP
+
 void lonlat2tile(double lon, double lat, int zoom, long long *x, long long *y);
 void epsg3857totile(double ix, double iy, int zoom, long long *x, long long *y);
 void tile2lonlat(long long x, long long y, int zoom, double *lon, double *lat);
@@ -15,3 +18,5 @@ struct projection {
 
 extern struct projection *projection;
 extern struct projection projections[];
+
+#endif