+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.21.0
+
+* Generate layer, feature, and attribute statistics as part of tileset metadata
+
+## 1.20.1
+
+* Close mbtiles file properly when there are no valid features in the input
+
+## 1.20.0
+
+* Add long options to tippecanoe-decode and tile-join. Add --quiet to tile-join.
+
 ## 1.19.3
 
 * Upgrade protozero to version 1.5.2